Rename ChordData file
